LPS-113730 Extract bundle refresh/install logic into osgi-util, prepare for reuse
