+++ modules/apps/static/osgi/osgi-util/bnd.bnd
@@ -1,6 +1,6 @@
 Bundle-Name: Liferay OSGi Utilities
 Bundle-SymbolicName: com.liferay.osgi.util
-Bundle-Version: 6.0.5
+Bundle-Version: 6.1.0
 Export-Package:\
 	com.liferay.osgi.util,\
 	com.liferay.osgi.util.bundle,\
