+++ modules/apps/static/osgi/osgi-util/src/main/resources/com/liferay/osgi/util/packageinfo
@@ -1,1 +1,1 @@
-version 2.1.0
+version 2.2.0
